requirements.txt
Flask==0.12.2
folium==0.5.0
python-dateutil==2.6.1